Add prefetching for project data
This commit adds prefetching for project data.

This is done by adding a `onMouseEnter` and `onFocus` handler to the project
links.

The handler will call `queryClient.prefetchQuery` for the project data, the
project users and the project invitations.

This will make the project page load faster when the user clicks on the link.