Add keyboard shortcuts to projects
This adds keyboard shortcuts to the projects page.

The shortcuts are based on the index of the project in the list.

For example, the first project in the list has the shortcut `1`, the second
project has the shortcut `2`, and so on.

The shortcuts are displayed using a ring around the project.