Add Pusher support
This adds support for Pusher to the NextAuth.js example app.

The Pusher client library is added to the client side code and the Pusher
server library is added to the server side code.

The Pusher server library is used to trigger events when a user signs in
or signs out.

The Pusher client library is used to listen for events and update the
user interface when a user signs in or signs out.

This is a breaking change because it adds a new dependency to the project.