Fix scrolling to bottom of chat
The previous implementation was scrolling to the bottom of the chat
window, but it was not taking into account the height of the chat
window. This meant that if the chat window was not full height, the
last message would be hidden.

This commit fixes the issue by scrolling to the bottom of the chat
window minus the height of the chat window.