Add a scroll to bottom feature to the chat
This is a feature that I've been wanting to add for a while.

I've been using the `scrollTo` method on the `ref` of the chat container.

I've also added a `shouldScroll` state variable to prevent the chat from scrolling to the bottom when the user is scrolling up.

I've also added a `setShouldScroll` function to set the `shouldScroll` state variable.

I've also added a `useEffect` hook to scroll to the bottom of the chat when the `shouldScroll` state variable is true.

I've also added a `setShouldScroll(false)` call to the `onSuccess` callback of the `useQuery` hook.

I've also added a `setShouldScroll(true)` call to the `useSubscribeToEvent` hook.